+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - AbstractValue.setAndNotify() bug fixed, change event no longer triggered when nullValue set.
 - Value.locked() added.
 - Value.changed() added, along with Value.Change.
+- Conditional no longer implements Consumer.
 ### is.codion.common.model
 - DefaultConditionModel, caseSensitive now included in locked state.
 ### is.codion.swing.common.model